Make all actions require an argument
Previously the default was to connect to a named pipe, with the option
of connecting to an assuan socket instead.

This changes the program to require an explict action to run - either to
connect to a named pipe or connect to an assuan socket.

The idea is that future connection types will be actions tacked on the
end, rather than yet more flags to mutate the default behaviour.